:art: grid popover chagned to dropdown
grid popover changed to dropdown
