+++ src/components/dash/workouts/month/GridDay.module.scss
@@ -1,0 +1,24 @@
+@import '../../../../styles/variables.scss';
+
+.dropdown {
+  width: 175px;
+  height: auto;
+  overflow: auto;
+  padding: 1rem;
+
+  .workout {
+    margin-bottom: 1rem;
+    width: 100%;
+    border-radius: 8px;
+    background: $primary;
+    color: white;
+    text-align: center;
+    font-size: 1.2rem;
+    padding: 0.5rem;
+    cursor: pointer;
+
+    &:hover {
+      text-decoration: underline;
+    }
+  }
+}
